Changed sell screen structure
Allows the image uploading for selling items to be more suitable.
